ai-server/requirements.txt
Flask
flask-cors
python-dotenv
google-generativeai
Gunicorn
json-repair